Changed type of index_type char/int to string.
The index_type is part of the RecWord structure which is used by all
record filters. It used to be a int char (e.g. 'w'). It is now a string
so that arbitrary index types can be specified . For example there might
be a Greek Title (e.g. word:greek or word_greek).