Allow record retrieval for record types using custom record IDs.
A filter may specify its own record_id by assigning a non-zero
to record_id member in RecWord. However, until this commit, the
original record was not retrievable . Updating only occured with
a custom match ID. With this patch the record can be retrieved .
The allrecords index (_ALLRECORDS) is used as a map between custom
record IDs and Zebras internal record store which is identified by the
system-number (storeData and storeKeys).