New attribute @fields for snippets that indicates match in other fields.
The snippets now has a new attribute @fields which includes one or
more field names of where the snippet *also* occur. This is useful
for any:w searches but one wants to know which other fields there is
identical terms. Note that this requires that the indexing uses
same sequence numbers . For DOM this means specifying the indexes
in the same <z:index directieve.