Fix bug WRT inconsistent registers, bug #3062.
For a many create+drop of DBs the ordinalSU will increase. This, in
turn will lead to a bigger resulting indexing term beyond length
IT_MAX_WORD - for the compressed temp files. This patch increases the
indexing buffer for the term to INP_NAME_MAX which is a log higher than
IT_MAX_WORD. The too-small size resulted in a truncations of long terms
(in fact phrases) that would be wrongly matched against the ISAM.