scan: Fix first-in-field term skipping
Different check for first-in-feild term. The check for first-in-field
term could skip CJK entries for ICU normalized sort keys. In theory it
can still happen if CJK entries would map to exactly the same sequence
as first-in-field entries (highly unlikely). There was also a problem
with scan for the empty string. Before this fix, it  would result in
scan failure (no entries returned), because first entry would hit the
the first-in-field term. Fixes bug #4207.